+++ src/lib/hooks/links/index.ts
@@ -1,2 +1,3 @@
 export * from './useHorizontalScroll';
 export * from './useFolderShare';
+export * from './useAddLink';
